Use separate build trees for development/release.
The motivation behind this was to definitiely identify which wasm file
is being used for profiling mode, because we want to be able to
identify which Rust function is being executed, instead of seing just
numbered WASM functions.

This change doesn't solve that problem, but it does mean that we know
a priori that the WASM file we're looking at is indeed the one build
for profiling mode.
