+++ tx2-web/built-files/.gitignore
@@ -1,0 +1,4 @@
+build
+development
+release
+profiling
