+++ tx2-web/built-files/README.md
@@ -1,0 +1,3 @@
+Generated files go here.
+
+This is controlled by scripts in ../package.json and by ../config.common.js.
