+++ tx2-web/src/controller/tx2.ts
@@ -1,4 +1,5 @@
-import { create_tx2, get_builtin_sample_tape, Tx2, tx2_codabo, tx2_device_statuses, tx2_do_tick, tx2_drain_device_changes, tx2_load_tape, tx2_lw_keyboard_click, tx2_next_simulated_tick, tx2_unmasked_alarm_active } from '../../build/tx2_web';
+import { create_tx2, get_builtin_sample_tape, Tx2, tx2_codabo, tx2_device_statuses, tx2_do_tick, tx2_drain_device_changes, tx2_load_tape, tx2_lw_keyboard_click, tx2_next_simulated_tick, tx2_unmasked_alarm_active } from '../../built-files/build/tx2_web';
+
 import { AlarmController } from './alarms'
 import { IoController } from './io'
 import { WasmUnitState } from './types'
