+++ tx2-web/webpack.development.js
@@ -1,0 +1,3 @@
+const config_common = require('./config.common.js');
+
+module.exports = config_common.make_webpack_config_exports("development");
